# nachweiseBackend/src/main/resources/api.yaml
openapi: "3.0.3"

info:
  description: API für Ausbildungsnachweise
  version: 1.0.0
  title: Ausbildungsnachweise

servers:
  - url: http://localhost:8080
    description: local

paths:
  /user:
    parameters:
      - $ref: '#/components/parameters/sessionId'
    post:
      operationId: postUser
      tags:
        - User
      description: create User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userCreate'
      responses:
        200:
          description: user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGet'
    patch:
      operationId: patchUser
      tags:
        - User
      description: Change User information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userPatch'
      responses:
        200:
          description: changed user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGet'

  /user/{userName}:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: userName
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getUser
      tags:
        - User
      description: get user information
      responses:
        200:
          description: user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userGet'

  /apprentice:
    parameters:
      - $ref: '#/components/parameters/sessionId'
    get:
      operationId: findUsers
      tags:
        - Apprentices
      responses:
        200:
          description: Get list of apprentices which proofs the logged in user ca see
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /apprentice/{userName}/requested:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: userName
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Apprentices
      description: get proofs requested for acknowledgement
      operationId: getRequestedForUser
      responses:
        200:
          description: list of proofs requested for acknowledgement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofPreview'

  /apprentice/{userName}/proofs:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: userName
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Apprentices
      description: get proofs which are acknowledged
      operationId: getForUser
      responses:
        200:
          description: list of proofs requested for acknowledgement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofPreview'

  /apprenticeship:
    parameters:
      - $ref: '#/components/parameters/sessionId'
    post:
      operationId: postApprenticeship
      tags:
        - Apprenticeship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/apprenticeshipData'
      responses:
        200:
          description: created apprenticeship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apprenticeshipGet'

  /apprenticeship/{id}:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      operationId: patchApprenticeship
      description: change ending date
      tags:
        - Apprenticeship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/apprenticeshipPatch'
      responses:
        200:
          description: changed ending date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apprenticeshipGet'

  /session/{username}:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/password'
      operationId: getSession
      parameters:
        - name: username
          in: path
          schema:
            type: string
          required: true
      tags:
        - Session
      responses:
        204:
          description: setting a session cookie
          headers:
            Set-Cookie:
              schema:
                type: string
                example: 'JSESSIONID=asdlfjkhasdfieu384=='

  /proof:
    parameters:
      - $ref: '#/components/parameters/sessionId'
    post:
      tags:
        - Proof
      description: Post a new proof
      operationId: postProof
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/proofPost'
      responses:
        200:
          description: Created Proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofGet'

  /proof/preview:
    parameters:
      - $ref: '#/components/parameters/sessionId'
    get:
      tags:
        - Proof
      description: get Proof previews
      operationId: getPreview
      responses:
        200:
          description: list of previews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofPreview'

  /proof/{id}:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: id
        in: path
        schema:
          type: integer
        required: true
    patch:
      tags:
        - Proof
      operationId: patchProof
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/proofPatch'
      responses:
        200:
          description: patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofGet'

    get:
      tags:
        - Proof
      operationId: getProof
      responses:
        200:
          description: patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/proofGet'

  /proof/{id}/acknowledgement:
    parameters:
      - $ref: '#/components/parameters/sessionId'
      - name: id
        in: path
        schema:
          type: integer
        required: true
    post:
      tags:
        - Acknowledgement
      operationId: postAcknowledgement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/acknowledgement'
      responses:
        200:
          description: "posted acknowledgement"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/acknowledgement'
    get:
      tags:
        - Acknowledgement
      description: Get List of messages and status
      operationId: getAcknowledgements
      responses:
        200:
          description: List of messages and status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/acknowledgementList'

components:
  parameters:
    sessionId:
      name: "JSESSIONID"
      in: cookie
      schema:
        type: string

  schemas:
    email:
      type: object
      properties:
        email:
          type: string
          example: 'email@example.com'

    password:
      type: object
      properties:
        password:
          type: string
          example: 'asd/8+_:#'

    userPatch:
      allOf:
        - $ref: '#/components/schemas/email'
        - $ref: '#/components/schemas/password'

    personalData:
      type: object
      properties:
        role:
          type: string
          enum:
            - 'Azubi'
            - 'Ausbilder'
            - 'Prüfer/Lehrer'
            - 'Ad'
        vorname:
          type: string
        nachname:
          type: string

    userCreate:
      allOf:
        - $ref: '#/components/schemas/userPatch'
        - $ref: '#/components/schemas/personalData'

    userData:
      allOf:
        - $ref: '#/components/schemas/personalData'
        - $ref: '#/components/schemas/email'
        - type: object
          properties:
            userName:
              type: string

    userGet:
      allOf:
        - $ref: '#/components/schemas/userData'
        - type: object
          properties:
            id:
              type: integer

    apprenticeshipData:
      type: object
      properties:
        start:
          type: string
          pattern: '[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])'
        end:
          type: string
          pattern: '[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])'
        apprenticeName:
          type: string
        instructorName:
          type: string
        profession:
          type: string

    apprenticeshipGet:
      allOf:
        - $ref: '#/components/schemas/apprenticeshipData'
        - type: object
          properties:
            id:
              type: integer

    apprenticeshipPatch:
      type: object
      properties:
        end:
          type: string
          pattern: '[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])'

    dayEntryData:
      type: array
      required:
        - workDone
        - hours
      items:
        type: object
        properties:
          workDone:
            type: string
          hours:
            type: integer

    dayGet:
      type: object
      required:
        - entries
        - department
        - hours
      properties:
        entries:
          $ref: '#/components/schemas/dayEntryData'
        department:
          type: string
        hours:
          type: integer

    dayPost:
      type: object
      required:
        - entries
        - department
      properties:
        entries:
          $ref: '#/components/schemas/dayEntryData'
        department:
          type: string

    dayPatch:
      type: object
      properties:
        entries:
          $ref: '#/components/schemas/dayEntryData'
        department:
          type: string

    proofPost:
      type: object
      required:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
      properties:
        monday:
          $ref: '#/components/schemas/dayPost'
        tuesday:
          $ref: '#/components/schemas/dayPost'
        wednesday:
          $ref: '#/components/schemas/dayPost'
        thursday:
          $ref: '#/components/schemas/dayPost'
        friday:
          $ref: '#/components/schemas/dayPost'
        saturday:
          $ref: '#/components/schemas/dayPost'
        requested:
          type: boolean
          default: false

    proofPatch:
      type: object
      properties:
        monday:
          $ref: '#/components/schemas/dayPatch'
        tuesday:
          $ref: '#/components/schemas/dayPatch'
        wednesday:
          $ref: '#/components/schemas/dayPatch'
        thursday:
          $ref: '#/components/schemas/dayPatch'
        friday:
          $ref: '#/components/schemas/dayPatch'
        saturday:
          $ref: '#/components/schemas/dayPatch'
        requested:
          type: boolean
          default: false

    proofGet:
      required:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - requested
        - weekStartDate
      type: object
      properties:
        monday:
          $ref: '#/components/schemas/dayGet'
        tuesday:
          $ref: '#/components/schemas/dayGet'
        wednesday:
          $ref: '#/components/schemas/dayGet'
        thursday:
          $ref: '#/components/schemas/dayGet'
        friday:
          $ref: '#/components/schemas/dayGet'
        saturday:
          $ref: '#/components/schemas/dayGet'
        acknowledgement:
          $ref: '#/components/schemas/acknowledgementList'
        requested:
          type: boolean
        weekStartDate:
          type: string
          pattern: '[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])'

    acknowledgement:
      type: object
      required:
        - acknowledged
        - message
      properties:
        acknowledged:
          type: boolean
        message:
          type: string

    acknowledgementList:
      type: object
      required:
        - acknowledged
        - messages
      properties:
        acknowledged:
          type: boolean
        messages:
          type: array
          items:
            type: string

    proofPreview:
      type: array
      items:
        type: object
        required:
          - id
          - weekStartDate
          - started
        properties:
          id:
            type: integer
          weekStartDate:
            type: string
            pattern: '[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[1-2][0-9]|3[0-1])'
          status:
            type: string
            enum:
              - SAVED
              - REQUESTED
              - REJECTED
              - ACCEPTED